LoRa, Accel, SD

+++ lib/measurements/bmp280.h
@@ -4,8 +4,8 @@
 
 typedef struct bmp280_data
 {
-    int32_t temperature;
-    int32_t pressure;
+    float temperature;
+    float pressure;
 } bmp280_data_t;
 
 void bmp280Init(int i2c, int sda, int scl);
